Consolidate checks into one loc
